Contacts: Cleanup API and Controller
