+++ src/server/controller/cycleData/contact/index.ts
@@ -1,0 +1,11 @@
+import { create } from './create'
+import { getMany } from './getMany'
+import { remove } from './remove'
+import { update } from './update'
+
+export const Contacts = {
+  create,
+  getMany,
+  remove,
+  update,
+}
